requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.8.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.8
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
Flask==3.0.3
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
importlib_metadata==8.5.0
iniconfig==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.1
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
openai==2.2.0
openpyxl==3.1.5
outcome==1.2.0
packaging==22.0
pandas==2.0.3
pdfminer.six==20231228
pdfplumber==0.11.5
Pillow==9.3.0
pluggy==1.0.0
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
PyMySQL==1.0.2
PyPDF2==3.0.1
pypdfium2==5.0.0
PySocks==1.7.1
pytest==7.2.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-rerunfailures==14.0
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
qrcode==7.3.1
regex==2024.11.6
requests==2.28.2
rich==14.3.1
selenium==4.27.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.46
sqlparse==0.4.3
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.67.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.20.1
typing_extensions==4.13.2
tzdata==2022.7
urllib3==1.26.14
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
diff-match-patch==20230430
pdf2docx==0.5.8
